Add centered y-coordinates for filtered data labels of vertical stacked bar charts

Calculates the y-coordinates for vertical stacked bar charts, centered for each group. The default function does not work if some labels are filtered.

add_label_position_typ_1(df, x, y)

Arguments

df
Data frame

x
x-coordinate

y
y-coordinate

Value

Data frame with additional column "label_" + the name of the y-coordinate variable

Examples

x <- c("WiSe 13/14", "WiSe 13/14", "WiSe 13/14", "WiSe 13/14")
y <- c(1989, 58, 163, 470)
fill <- c("Bachelor 2-Fächer", "Master 1-Fach", "Master 2-Fächer", "Master of Education")
filter_cutoff <- 0.05
df <- get_label_position(df, x, y)

#> Error in get_label_position(df, x, y): konnte Funktion "get_label_position" nicht finden

ggplot2::ggplot() +
  ggplot2::geom_bar(aes(x = x, y = y, fill = fill), stat = "identity") +
  ggplot2::geom_label(data = df, aes(x = x, y = y_label, group = fill, label = y))

#> Error in aes(x = x, y = y, fill = fill): konnte Funktion "aes" nicht finden