Plot vertical stacked bar charts that are scaled to 100\ rub_plot_type_2( df, x_var, y_var, y_axis_label = "", fill_var, fill_label = NULL, fill_reverse = FALSE, facet_var = NULL, caption = "", caption_prefix = "Quelle:", filter_cutoff = 0.04, color = RUB_colors["blue"], palette_reverse = FALSE, base_family = "RubFlama", base_size = 11, max_width_strip_label = 80 ) dfData frame x_varRequired variable name for the variable containing the discrete x-coordinates. y_varRequired variable name for the variable containing the continuous y-coordinates. y_axis_labelOptional label for the y-axis, defaults to an empty string. fill_varVariable name for the discrete variable which determines the groups to be stacked, e.g. degree. fill_labelOptional variable name for the character variable containing the names of the fill variable, defaults to NULL. fill_reverseBoolean indicating whether the order of the fill variable should be reversed, defaults to FALSE. facet_varOptional variable name for the discrete variable to facet by, defaults to NULL. captionOptional character containing the data source for the figure (prefix 'Quelle:' is automatically added). caption_prefixOptional character containing the prefix for the caption, defaults to 'Quelle:'. filter_cutoffOptional cutoff value for the suppression of data labels. By default, all values below 0.04 of the total value of the stacked bar chart are suppressed. colorColor for font and borders, defaults to RUB_colors["blue"], i.e. #003560. palette_reverseOptional boolean indicating whether the colors in the palette should be reversed, defaults to FALSE. base_familybase font family, defaults to RubFlama base_sizebase font size, defaults to 11 max_width_strip_labelOptional maximum width in characters for the facet label passed to ggplot2::label_wrap_gen. A ggplot object Plot vertical stacked bar charts that are scaled to 100\ # Create test values for all three mandatory variables (x_var, y_var, fill_var) df_t2_ex1 <- tibble::tribble( ~cohort_term, ~status_percentage, ~cohort_status, "2. cohort term", 0.9513551740, "Studying", "2. cohort term", 0.0029748098, "Changed subject", "2. cohort term", 0.0004673679, "Graduated", "2. cohort term", 0.0186648938, "Disenrolled without degree", "2. cohort term", 0.0265377545, "Dropped subject", "4. cohort term", 0.8896149868, "Studying", "4. cohort term", 0.0616919929, "Changed subject", "4. cohort term", 0.0016484686, "Graduated", "4. cohort term", 0.0201024499, "Disenrolled without degree", "4. cohort term", 0.0269421019, "Dropped subject", "6. cohort term", 0.7901183540, "Studying", "6. cohort term", 0.1502641318, "Changed subject", "6. cohort term", 0.0074548056, "Graduated", "6. cohort term", 0.0243490259, "Disenrolled without degree", "6. cohort term", 0.0278136827, "Dropped subject", "8. cohort term", 0.6115873010, "Studying", "8. cohort term", 0.2961468339, "Changed subject", "8. cohort term", 0.0104080044, "Graduated", "8. cohort term", 0.0274549015, "Disenrolled without degree", "8. cohort term", 0.0544029593, "Dropped subject", ) rub_plot_type_2( df = df_t2_ex1, x_var = cohort_term, y_var = status_percentage, fill_var = cohort_status )