Get formatted flextable of response rates for the Eingangsbefragung

Get formatted flextable of response rates for the Eingangsbefragung

Usage

rub_table_eb(df, typology, headings)

Arguments

df

Data frame

typology

Data frame with flextable typology

headings

Character vectors of headings

Value

Formatted Flextable

Examples

# Generate example data
df_example <- structure(
  list(
    studieneingang = c("Geschlecht", "weiblich", "männlich", "Abschlussart", "Bachelor 1-Fach", "Bachelor 2-Fächer", "Magister Theologiae", "Master 1-Fach", "Staatsexamen", "Fächergruppe (erstes Studienfach)", "Geisteswissenschaften", "Humanmedizin / Gesundheitswissenschaften", "Ingenieurwissenschaften", "Kunst, Kunstwissenschaft", "Mathematik, Naturwissenschaften", "Rechts-, Wirtschafts- und Sozialwissenschaften", "Sport", "Gesamtzahl angeschriebene Studierende / gültige Fragebögen", "Rücklaufquote Studieneingangsbefragungen"),
    koepfe_rub = c(NA, 7513, 7464, NA, 8150, 3615, 54, 889, 2260, NA, 4006, 886, 2864, 243, 3574, 3136, 266, 14975, 0.31),
    koepfe_rub_perc = c(NA, 0.5, 0.5, NA, 0.54, 0.24, 0, 0.06, 0.15, NA, 0.27, 0.06, 0.19, 0.02, 0.24, 0.21, 0.02, 1, 0.31),
    koepfe_bef = c(NA, 2826, 1881, NA, 2284, 1234, 18, 303, 864, NA, 1407, 416, 688, 54, 1116, 921, 105, 4707, 0.31),
    koepfe_bef_perc = c(NA, 0.6, 0.4, NA, 0.49, 0.26, 0, 0.06, 0.18, NA, 0.3, 0.09, 0.15, 0.01, 0.24, 0.2, 0.02, 1, 0.31),
    row_id = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19)
  ),
  class = c("data.frame"),
  row.names = c(NA,-19L)
)

# Multi-level headings, see `flextable::set_headers`
typology_example <- structure(
  list(
    col_keys = c("studieneingang", "koepfe_rub", "koepfe_rub_perc", "koepfe_bef", "koepfe_bef_perc"),
    colC = c("Studieneingang", "Studierende im 1. HS (WiSe 18/19 bis  WiSe 20/21)", "Studierende im 1. HS (WiSe 18/19 bis  WiSe 20/21)", "Studierende im 1. HS (WiSe 18/19 bis  WiSe 20/21)", "Studierende im 1. HS (WiSe 18/19 bis  WiSe 20/21)"),
    colB = c("Studieneingang", "Angeschrieben", "Angeschrieben", "Gültige Fragebögen", "Gültige Fragebögen"),
    colA = c("Studieneingang", "Köpfe", "(in %)", "Köpfe", "(in %)")
  ),
  class = c("data.frame"),
  row.names = c(NA, -5L)
)

# Text for rows that receive special formatting
headings_example <- c("Geschlecht", "Abschlussart", "Fächergruppe (erstes Studienfach)", "Gesamtzahl angeschriebene Studierende / gültige Fragebögen", "Rücklaufquote Studieneingangsbefragungen")

# Function call
rub_table_eb(
  df = df_example,
  typology = typology_example,
  headings = headings_example
)
#> a flextable object.
#> col_keys: `studieneingang`, `koepfe_rub`, `koepfe_rub_perc`, `koepfe_bef`, `koepfe_bef_perc` 
#> header has 3 row(s) 
#> body has 19 row(s) 
#> original dataset sample: 
#>    studieneingang koepfe_rub koepfe_rub_perc koepfe_bef koepfe_bef_perc row_id
#> 1      Geschlecht         NA              NA         NA              NA      1
#> 2        weiblich       7513            0.50       2826            0.60      2
#> 3        männlich       7464            0.50       1881            0.40      3
#> 4    Abschlussart         NA              NA         NA              NA      4
#> 5 Bachelor 1-Fach       8150            0.54       2284            0.49      5